route safety + listing ownership by user

+++ routes/web.php
@@ -22,32 +22,40 @@ Route::get('/', [ListingController::class, 'index']);
 Route::post('/listings/', [ListingController::class, 'store']);
 
 // show create listing form
-Route::get('/listings/create', [ListingController::class, 'create']);
+// middleware for user owned listings
+// auth middlewear for route protection
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 // show listing edit form
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+// auth middlewear for route protection
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 // update listing
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+// auth middlewear for route protection
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 // delete listing
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+// auth middlewear for route protection
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
 // show individual listing by id
 // should be declared last, to avoid path conflict after /listings/
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
 // show new user register form
-Route::get('/register', [UserController::class, 'create']);
+// guest middlewear for route protection
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
 // create new user
 Route::post('/users', [UserController::class, 'store']);
 
 // logout user
-Route::post('/logout', [UserController::class, 'logout']);
+// auth middlewear for route protection
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
 
 // show login form
-Route::get('/login', [UserController::class, 'login']);
+// guest middlewear for route protection
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
 
 // login user
 Route::post('/users/authenticate', [UserController::class, 'authenticate']);
